added click event to all tree coconuts. refactored name of coconut component to coconutbutton for better specificity.
