+++ src/components/TreeCoconut.js
@@ -1,25 +1,66 @@
 
 import React, { Component } from 'react';
-import { View, StyleSheet, Image } from 'react-native';
+import { View, StyleSheet, Image, TouchableOpacity } from 'react-native';
 import coconut from '../img/coconut.png'
 
 class TreeCoconut extends Component {
-  
+
+  handleClickedCoconut = () => {
+    console.log("i clicked on the event lister")
+  }
+
+
   render(){
     return(
       <View style={styles.coconutContainer}>
-        <Image source={coconut} style={styles.coconutBunch1A} />
-        <Image source={coconut} style={styles.coconutBunch1B} />
-        <Image source={coconut} style={styles.coconutBunch1C} />
-        <Image source={coconut} style={styles.coconutBunch2A} />
-        <Image source={coconut} style={styles.coconutBunch2B} />
-        <Image source={coconut} style={styles.coconutBunch2C} />
-        <Image source={coconut} style={styles.coconutBunch3A} />
-        <Image source={coconut} style={styles.coconutBunch3B} />
-        <Image source={coconut} style={styles.coconutBunch3C} />
-        <Image source={coconut} style={styles.coconutBunch4A} />
-        <Image source={coconut} style={styles.coconutBunch4B} />
-        <Image source={coconut} style={styles.coconutBunch4C} />
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch1A} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch1B} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch1C} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch2A} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch2B} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch2C} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch3A} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch3B} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch3C} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch4A} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch4B} />
+          </TouchableOpacity>
+          <TouchableOpacity onPress={ ()=> 
+            this.handleClickedCoconut()} >
+            <Image source={coconut} style={styles.coconutBunch4C} />
+          </TouchableOpacity>
       </View>
     )
   }
